+++ src/helpers/__tests__/display.js
@@ -1,0 +1,11 @@
+import display from "../display";
+
+it("should have all the display options", () => {
+  const displays = ["none", "block", "inline-block", "inline", "table", "flex"];
+
+  displays.forEach(key => {
+    expect(display[key].resolve()).toEqual({
+      display: key,
+    });
+  });
+});