+++ src/helpers/__tests__/round.js
@@ -1,0 +1,8 @@
+import round from "../round";
+
+it("has a round helper", () => {
+  expect(round.resolve()).toEqual({
+    borderRadius: "50%",
+    content: '""',
+  });
+});