+++ src/helpers/hidden.js
@@ -1,0 +1,12 @@
+// @flow
+import Style from "further";
+
+export default Style.of({
+  visibility: "hidden",
+  height: 0,
+  maxHeight: 0,
+  margin: 0,
+  paddingTop: 0,
+  paddingBottom: 0,
+  opacity: 0,
+});