+++ src/helpers/round.js
@@ -1,0 +1,8 @@
+//@flow
+
+import Style from "further";
+
+export default Style.of({
+  borderRadius: "50%",
+  content: '""',
+});